refactor(web): extract AndroidPttFields component
